"zero" and "sıfır" bugs fixed -> Stable Version

+++ TranslatorService/src/translatorservice/impl/ITranslator.java
@@ -3,6 +3,6 @@ package translatorservice.impl;
 public interface ITranslator {
 	
 	public String numberToWord(int number, String locale);
-	public int wordToNumber(String word, String locale);
+	public double wordToNumber(String word, String locale);
 
 }
